Patches to remove problematic string collections: (r=daumling)
-- Force String::getData() to go thru a stack-based temporary (StringData) so that the intermediate strings for concats, etc can be found by GC and not collected.

-- restructure the append() method to pass the string further down the chain; due to tail-call optimization in some compilers the upstream stack reference was being lost and allowed GC to collect a string in use.

-- make String::toUTF8String() and String::toUTF16String private; now you must use StUTF8/16String to access these, to ensure there's a valid stack reference to the UTF string in question.

-- fixed an unrelated 64-bit bug in Sampler.cpp that was crashing SimpleSampling.abc

+++ core/Sampler.cpp
@@ -275,8 +275,7 @@ namespace avmplus
 #ifdef DEBUGGER					
 		if(typeOrVTable < 7 && core->codeContext() && core->codeContext()->domainEnv()) {
 			// and in toplevel
-			// FIXME 64bit
-			typeOrVTable |= (uint32)(uintptr_t)core->codeContext()->domainEnv()->toplevel();
+			typeOrVTable |= (uintptr)core->codeContext()->domainEnv()->toplevel();
 		}
 #endif
 		AvmAssertMsg(s.sampleType == NEW_AUX_SAMPLE, "Sample stream corrupt - can only add info to an AUX sample.\n");
